bazel syntax: replace StarlarkThread.eventHandler by PrintHandler
The sole purpose of the thread's EventHandler was for use by the
built-in "print" function. This change replaces it by a specialized
interface, PrintHandler, enabling implementations to do richer
reporting, for example including the complete call stack at the moment
of a debug print statement.

StarlarkThread.setPrintHandler updates the thread's print handler.
We do not add it to the Builder, as StarlarkThread really doesn't need
a builder: it is the very opposite of an immutable object, it is
accessed only by a single thread, and most of the builder fields
don't belong in StarlarkThread anyway.

StarlarkThread.makeDebugPrintHandler creates an adaptor from Starlark
print events to Bazel debug events. This function will move to lib.events
when we reverse the syntax<->events dependency.

PiperOrigin-RevId: 288762547