Objc: avoid implicit iteration over nested set
This change should only have "obviously correct" transformations of
the source code to avoid introducing any regressions and to make it
easier for future maintainers of this code to see which variables /
fields refer to what.

This does not cover all uses of NestedSet; especially the
CompilationArtifacts class uses Iterables.concat extensively, which
currently does *not* flatten the NestedSet, even though the
documentation says that they are flattened. AFAICT, they are flattened
on read, not on write, which is the worst of both worlds (if there are
multiple reads to the same object).

PiperOrigin-RevId: 288880929