Put HTTP portions of external downloads behind an interface.
This PR is in preparation for a remote implementation of external downloads. It is intended to have no behavior changes, and to split the current `HttpDownloader` logic into:
* `HttpDownloader`, which does the actual download with all the fancy concurrency and error handling.
* `DownloadManager`, which interacts with `RepositoryCache` and the distdirs. It doesn't know about HTTP any more, instead going through a `Downloader` interface.

A future PR will implement a `Downloader` in terms of the bazelbuild/remote-apis protocol and put it behind an `--experimental_remote_downloader` flag.

R: @buchgr @dslomov
CC: @sstriker

Closes #10547.

PiperOrigin-RevId: 288884556