BuildView refactor: collectTests is now static
I renamed scheduleTestsIfRequested method to
collectTests, because the method does not schedule
any tests but just separates the parallel tests
from the exclusive ones.

I also changed the method's signature to return
the two sets rather than mutate arguments, because
this makes the method pure (free of side effects)
so it is easier to reason about.

PiperOrigin-RevId: 288900132