Make dynamic scheduler aware of local-only execution info, and optionally fail builds with dynamic execution if the Xcode related execution requirements haven't been set.
This follows the modification of (hopefully all) Xcode-related rules to set action execution requirements based on Xcode version. Right now, this should not change behavior, both because the flag that enables the failure defaults to false and because the Xcode config that sets the execution requirements is not yet being used (so we don't expect any actions to have different functional execution requirements yet). However, turning on this failure (via --experimental_require_availability_info) will be used to ensure that we've changed all of the Xcode related rules, which is important because builds could fail cryptically if actions are run on the wrong platform.

RELNOTES: None.
PiperOrigin-RevId: 288947386