FileType.getExtensions() now returns ImmutableList
Change the signature to return ImmutableList
instead of List, because this clearly signals to
callers that it is not a mutable list, and because
every implementation was already returning an
immutable list anyway.

This change allows avoiding some
ImmutableList.copyOf calls (with arguments that
are already immutable), so we create less garbage.

PiperOrigin-RevId: 289095325