Partial roll-forward of https://github.com/bazelbuild/bazel/commit/9700ee6778706e6be65c377cb13f758e18f4a3ae of the test runner changes.
The culprit of the rollback was an additional java_proto dependency on the test runner (which is used from head, not built at the same time with bazel). The dependency comes from using the worker protocol to communicate between bazel and the test runner.

The roll-forward is fixing the culprit by using jarjar to rename all protobuf and worker protocol's classes. Additionally remove everything else in the persistent test runner jar to avoid one version errors.

RELNOTES: None.
PiperOrigin-RevId: 289239195