C++: Add dynamic_deps attribute to cc_test
This was already added to cc_binary. This CL just adds the attribute in the rule definition of cc_test, the code path is the same as for cc_binary.

All of this is still guarded behind --experimental_cc_shared_library. Since the definition of cc_shared_library is in the external rules_cc repository and while this is experimental we don't want to couple changes to the bzl file with Bazel releases, we have the tests in @rules_cc which has an integration test using Bazel head. Because of that we first have to submit this CL, then the tests in unknown commit.

Rolling forward after depot fix in unknown commit

RELNOTES:none
PiperOrigin-RevId: 289390298