RepositoryResolvedEvent: be lazy about computing the directory hash
The RepositoryResolvedEvent is generated each time a respository was succesfully
fetched. However, not in all bazel invocations, the resolved object is actually
used (as not in all cases a resolved file is written). Therefore, delay the expensive
part of this operation, the computation of the hash of the generated directory,
till it is needed for the first time.

Closes #10147

Change-Id: I9cd9576d75ac8d6aa57ff8f9572417ace14b385a
PiperOrigin-RevId: 289449062