Move ActionContext resolution out of dynamic strategy.
Previously the dynamic strategy implementation would inspect all registered strategies (and other action contexts) to determine the ones corresponding to the strategy identifiers passed for the local and remote branch of dynamic execution. In an effort to encapsulate all identifier to strategy resolution in one place this work is now being performed in SpawnActionContextMaps, with a generic interface available for configuring the mapping and obtaining the results.

As part of this change the dynamic strategy tests now encompass the logic in DynamicExecutionModule which plays a central part in configuring the dynamic strategy.

RELNOTES: None.
PiperOrigin-RevId: 289462410