Move remote local fallback resolution to registry.
Previously the remote local fallback logic needed to inspect all registered strategy (and action context) implementations to find one that matches the (potentially) registered strategy commandline identifier for this purpose. In an effort to limit strategy resolution to one code location this change instead introduces a generic way of registering the user-provided commandline identifier and an API to retrieve the resolved strategy implementation.

RELNOTES: None
PiperOrigin-RevId: 289473807