Make tests not special when registering strategies.
Normally strategies and action contexts are registered with Bazel's execution phase through a builder which is populated by modules. Not so for --test_strategy: This flag's value is passed directly to the construction of strategy objects. This change removes this special treatment.

Note that this means that it is no longer guaranteed that a test strategy is registered with Bazel in all cases (although in practice it is). This is intentional as it is unclear why the test strategy should have this special requirement (in fact many commands don't run tests!).

RELNOTES: None
PiperOrigin-RevId: 289475025