Remove special-casing of enum/flag ID resources
aapt2 already emits the enums/flags as separate ID resources, which go through
the normal resource merging logic.

Enum/flag attributes are added as IDs to resources.arsc as an implementation
detail to support linking XML against another APK, and the fact that R.id.*
fields are generated is an unintended side effect which no one should rely
on.  RPBB goes out of its way to support this, making it harder to enforce
rules on resource definitions.

This allows changing a Map#merge to a Map#put, which is a 2x speedup for a
SortedMap.

PiperOrigin-RevId: 289523255