cquery now handles 'aliased configurations'
In Skyframe, a BuildConfiguration (in BuildConfigurationValue) can actually be aliased to multiple SkyKeys. The only current known aliasing is when the FragmentClassSet of the SkyKey contains 'extra' entries that will not result in creating Fragments. (e.g. TestConfiguration will only be built if TestOptions is in BuildOptions.)

ConfiguredTargetQueryEnvironment::getTransitiveConfigurations did not handle this properly by assuming all SkyKeys pointed to a unique BuildConfiguration. This fixes that issue by telling ImmutableMap to 'merge' the equal configurations.

Note: These aliased BuildConfiguration are value equal but not referentially equal so there is some memory waste here. Also, still susceptible to crashes via BuildConfiguration checksum collision; however, that would be its own crisis.

Testing Note: The test is a little awkward and relies on secret knowledge that dropping TestOptions causes TestConfiguration to be dropped from fragments.

RELNOTES: None.
PiperOrigin-RevId: 289533776