Add plumbing to allow querying all resources referenced from (protobuf-)XML
These are needed to validate resource references, i.e. references should be to
internal definitions or public resources from direct dependencies.

This entails adding a field for the protos to android/xml/*XmlResourceValue,
which is technically redundant to the existing strings/maps.  The latter can
be removed once we stop parsing XML---presently the XML parsing is only done
in AAR generation for the benefit of merging res/ directories.

PiperOrigin-RevId: 289551860