Refactor: remove unnecessary ImmutableXXX.copyOf()
It's often a code smell if a constructor makes an
immutable copy of its argument, because often
those arguments could easily be an immutable
collection.

In this CL I mark arguments of such ctors as
Immutable and let the caller make the copy that
the ctor used to.

This CL doesn't introduce any new copies, because
it merely moves the copy calls to the call sites
(where copying is needed at all).

PiperOrigin-RevId: 289612603