Change analysis_v2.proto ActionGraphContainer to the original format.
The main motivation for the previous change in proto format was to allow
streaming of the result. We made use of the oneof syntax to achieve this.
However, this would require the users of the proto to do an additional check to
retrieve the actual message behind the oneof wrapper.

This is undesirable as it makes the proto harder to use. This CL reverts that,
and instead keeps the old ActionGraphContainer in analysis.proto, where no
oneof check is required for users.

RELNOTES: None
PiperOrigin-RevId: 289628674