bazel syntax: remove the toplevel environment hack
Currently, the value of CallFrame.locals (formerly
StarlarkThread.lexicalFrame) for the <toplevel> function is an alias
for StarlarkThread.globalFrame, which is the Module of the <toplevel>
function, necessitating various hacks.

This change causes every function, including top-level ones such as
<toplevel> and <expr>, to have a local environment block.
Variables are bound in this block by load statements and by
comprehensions. We cannot yet assume that the variables bound
by load and by comprehensions have been properly resolved, so
we (in effect) do the resolution to Local during evaluation.
(This requires keeping track of whether evaluation is within a comprehension.)

This fixes issue https://github.com/bazelbuild/starlark/issues/92

Now that CallFrame.locals is never an alias for Module, we can
specialize its type from Frame to LexicalFrame, then inline that type
completely and throw away Frame. CallFrame.locals is now just a map.
Updating it does not throw MutabilityException.

Also, inline Eval.updateAndExport.

PiperOrigin-RevId: 289668675