Improve WORKSPACE file generation in shell tests.
By passing in the workspace name to create_workspace_with_default_repos, we can remove custom code for this from a few tests. This also ensures that the tests get a fresh WORKSPACE file with known contents when they call that method. The old code simply appended a few repositories to any existing WORKSPACE file which caused problems when overriding repositories.

RELNOTES: None.
PiperOrigin-RevId: 289830653