"Stream" action cache digest computations.
Now that ActionAnalysisMetadata.getInputs() returns a NestedSet, we don't need allocate an intermediate (path -> metadata) map during action digest computation. Encapsulate the logic for this in a OrderIndependentHasher class. This class also caches fingerprints and digest arrays to avoid allocations.

Update the size constant in CompactPersistentActionCacheTest.testToStringIsntToBig because the digests are all 32 byte arrays instead of 1 byte arrays.

Additionally, fix ActionCache.Entry so it's stateless and immutable by making a builder class for it. Delete CompactPersistentActionCacheTest.testEntryToStringIsIdempotent because the property it wants to proven by the type system.

Closes #10537.

PiperOrigin-RevId: 289858351