Evaluator: delete this unused interface
Nothing implements the nested Factory interface so
that can be deleted.

Only ParallelEvaluator implements Evaluator so the
interface can be deleted and the few references to
Evaluator updated.

I seem to remember that there used to be various
Evaluator implementations, but evidently these are
no more.

We could rename ParallelEvaluator to just
Evaluator but that'd touch more than 50 files,
which would be too much effort for negligible
benefit.

RELNOTES: none
PiperOrigin-RevId: 289872590