Replace @ExecutionStrategy with module-based registration.
Strategies and other action contexts used an annotation to indicate their identifying type (for example: "I'm something that can write a file for an action") and commandline identifiers (for example: "local", "remote"). This approach has a number of drawbacks:

* Reflection is required to obtain the values, making it harder to understand control flow/dependencies.
* Classes have to be created just so an annotation can be put upon them (everything else is handled as constructor parameters passed to a common base class).
* Identifying information is not visible in places usually associated with configuration (Bazel modules) which merely register the class.

This change removes the annotation-based propagation of identifying information and instead adds registration methods (called in Bazel modules that would previously simply register an annotation context instance) that take the data explicitly.

RELNOTES: None
PiperOrigin-RevId: 289881876