Fix linux sandbox memory leak
getcwd result is allocated using malloc and should be deallocated with
free.

However opt.working_dir won't own it because that's an std::string and
the current operator= call will allocate a new buffer and copy the
content of getcwd result.

Hence, in the previous solution, the result of getcwd was just leaked at
the end of the function.

This commit adds the necessary free.