Extract ActionContextRegistry interface for types that allow action context lookup.
Several types in Blaze offer a way to look up action context instances by their "identifying type", each along with a number of other purposes. This change unifies this behavior in a single interface so that the implementations can be used interchangeably and (where feasible) less abstraction leak is possible by just exposing this interface to consumers.

RELNOTES: None.
PiperOrigin-RevId: 289947907