Don't expose used contexts for searching, instead use a callback.
A context (or strategy) implementation is considered "used" if it is not only registered but also reachable from at least one mnemonic or set as a default. Some logic requires this information, for example to conditionally do some setup/cleanup if one of a set of strategies ended up being used.

To localize the impact of the "used information propagation" this change removes the global way to inspect all registered contexts. Instead it introduces a callback mechanism where each context gets notified if it is used and can then invoke a callback which is listened to by the same logic that previously inspected all contexts. This approach can further be simplified in the future when the explicit notification is no longer necessary (contexts/strategies will only be instantiated if used, giving an implicit callback location).

This is a precondition for disentangling context registration and context lifecycle management which will be concluded in a future change.

RELNOTES: None
PiperOrigin-RevId: 290055330