Add experimental_do_not_use_dont_symlink_in_execroot workspace function
enabled with the flag --experimental_ninja_actions.

Additionally, construct WorkspaceFactory with StarlarkSemantics attribute,
and use it for gathering workspace functions.
The reason for this change is that some workspace functions may be guarded by flags, so the usage of default StarlarkSemantics for gathering available workspace functions is not correct.
Without this change, the test WorkspaceFunctionTest of introduced workspace function, guarded by a flag, did not work.

Closes #10589.

PiperOrigin-RevId: 290277001