bazel sync: add an option to only sync a particular set of repositories
Add an option --only to `bazel sync` allowing to specify that only a specified
set of external repositories be synced. Still, all repositories (or only the
configure-like ones, if --configure is given) are considered out of date; so if
the specified repository depends one of those, those are fetched as well.

Change-Id: I1712963077081452fbaca0f34e2ab9415d80f880
PiperOrigin-RevId: 290583672