Rewrite the inprocess symlink strategy
This is currently hidden behind an experimental flag, so this code is
not used yet. This new implementation more closely matches
build-runfiles.cc. Also, it now correctly creates the output manifest,
which was previously missing.

This passes almost all of our tests, but there are still some open issues:
- on Windows, it copies instead of symlinking (this is caught by tests)
- it does not fix the file mode bits if there is an existing file or
  directory with incorrect mode bits (not caught by tests)

We are also not sure whether this is acceptable from a performance
standpoint.

PiperOrigin-RevId: 290944709