Artifact conflict check: remove unnecessary check
Don't check for artifact conflicts if it's a null
build and the only change since the last build is
that some SkyKeys were deleted.

Deleting ConfiguredTargets can only ever resolve
and not create conflicts:
- if the last build had a conflict, then we'd
  re-check anyway, because that is the same case
  as when building (x y) is in conflict but then
  null-building just (x) is not, and we re-check
  solely on the basis of the last build having had
  conflicts.
- if the last build had no conflict, then deleting
  a CT won't change that.

This CL also adds a comment that explains why we
don't check in this case.

This is a follow-up to commit 044264ee7290292195d0980657469ac69ba280e0.

RELNOTES: none
PiperOrigin-RevId: 291111167