Let http_archive copy WORKSPACE and BUILD files instead of symlinking them.
When using the "workspace_file" or "build_file" attributes of e.g. http_archive, we symlinked the files into the external repository. However, this means that they cannot be patched via the "patch_cmds" feature, because then the patch would modify the original source files.

Copying seems fine, because these files are usually small (I would argue that if a WORKSPACE or BUILD file is so huge that copying vs. symlinking it here makes a performance difference, then something is very wrong with these files). On Windows, we used copies anyway, because symlinking is weird. So, this change just normalizes the behavior over all platforms.

RELNOTES: None.
PiperOrigin-RevId: 291144287