Query performance: Skip aspect handling if none
Don't collect or try to convert aspect information if the rule class
doesn't have any aspects. To that end, keep track of which rule classes
have aspects and which don't.

It makes only a small difference in performance, but I think it's worth
it. I measure an improvement of a 2-3% on a query with a large proto
output.

PiperOrigin-RevId: 291144350