Python coverage: experimental first steps
- Add lcov_merger as an attribute to the py_test rule so that it does
  not immediately fall over
- Add a small section to python_stub_template.txt to perform minimal
  setup for coverage integration

I was able to get this to work manually by using a modified version of
coverage.py, see:
https://github.com/nedbat/coveragepy/pull/863

Note that the patch does not work out of the box.

It's basically ~impossible to add integration tests until we get a
working version of coverage.py, but this is the minimum change required
in Bazel to get a basic working version.

Closes #10433.

PiperOrigin-RevId: 291166690