Fix test_tar_utf8 to pass on macOS Catalina.
For whatever reason, the umlauts when encoded in NFC form were not recognized as valid UTF-8 by macOS Catalina when extracted by Bazel's tar archive extractor. When encoding them in a slightly different form (NFD), it all works.

I'm keeping the NFC form commented out so that we can revisit this in the future, but for now this test is blocking us from upgrading the CI machines to macOS Catalina.

RELNOTES: None.
PiperOrigin-RevId: 291168862