execution transition should no longer reset trim_test_configuration
As explained in the bug, the execution transition was improperly resetting trim_test_configuration causing unnecessary creation of ConfiguredTargets (which was also leading to functional issues as duplicate CppActions are not de-duplicated).

Fix by causing execution transition to preserve the trim_test_configuration flag.

Refactored tests a bit. Renamed the old TestTrimmingTransitionTest to TrimTestConfigurationTest as it was doing build-level tests (with fake bzl and BUILD files). The new TestTrimmingTransitionTest focuses on directly testing the TestTrimmingTransition class.

RELNOTES: None
PiperOrigin-RevId: 291173657