bazel syntax: fix crash caused by interrupting ParamDescriptor.evalDefault
When a Skyframe thread is interrupted, Starlark evaluation is supposed to
terminate quickly. However, if Starlark evaluation is in the middle of a
call to the method cache, then the cache population task, which recursively
calls the Starlark interpreter to parse parameter default value expressions, is
interrupted too, causing it to crash because it was not expecting interrupts.

This change causes the recursive Starlark evaluation to ignore java.lang.Thread
interrupts so that we can safely assume that they do not occur.

LocalCache should really use a pool thread for its entry computation tasks.
Also, Thread.interrupted and indeed thread-local storage should be excluded
from future designs for programming languages.

Thanks to Mark Schaller and Nathan Harmata for helpful design discussions.

PiperOrigin-RevId: 291250699