Add s390x and make zip -d step more resilient
Continuation of PR #10643 for application to third_party directory

Build on s390x fails because the process to remove unnecessary libraries from netty was not configured for x86. As a result, zip -d failed because it had nothing to remove. This PR does two things:

Adds conditions for s390x to the src and third_party BUILD files, and
Adds a condition to the genrule so that 'zip -d' does not execute if there's nothing to do.

See also Issue #9263; Pull Requests #9346 and #9945

Closes #10798.

Signed-off-by: Philipp Wollermann <philwo@google.com>