Implement changes in symlink forest creation related to dont_symlink_directories_in_execroot.
This is a continuation of https://github.com/bazelbuild/bazel/pull/10589.

If dont_symlink_directories_in_execroot is used in WORKSPACE file with some directories (immediate children of the workspace root directory), do not symlink those directories under the execroot during symlink forest creation.
(See SymlinkForest class for more context, and also https://docs.bazel.build/versions/master/output_directories.html)

Normally, source directories are symlinked to the execroot, so that the actions can access the input (source) files.
In the case of Ninja execution (enabled with --experimental_ninja_actions flag), it is typical that the directory with build-related files contains source files for the build, and Ninja prescribes creation of the outputs in that same directory.
Since commands in the Ninja file use relative paths to address source files and directories, we must recreate the exact same directory structure under the execroot.

However, Bazel requires output directories to be separated from input directories: it is Bazel?s concept, not something that is related to this change or Ninja integration.

That is why Ninja build configuration directory (though it is part of the sources in the sense that it contains the source files) can not be directly symlinked to the execroot (if it was, the source tree would be changed by the build).

dont_symlink_directories_in_execroot can be used to specify that Ninja build configuration directories should not be symlinked to the execroot.
It is not expected that there could be other use cases for using this method.

Closes #10600.

PiperOrigin-RevId: 291681522