Fix race when extracting the install base when Bazel runs in a PID namespace.
The PID is not a unique identifier in that case and if multiple instances of Bazel tried to extract itself to a shared location at the same time, they occasionally corrupted their temporary extraction directories. This is sometimes the case on CI systems (most notably Bazel's own CI), because a shared install base reduces the overhead of running the integration test suite a lot, as otherwise every test will extract its individual copy of Bazel.

RELNOTES: None.
PiperOrigin-RevId: 291685865