MacOS C++ coverage: preserve GCOV_PREFIX_STRIP
On MacOS, setting GCOV_PREFIX_STRIP to 3 is generally incorrect. This
setting works on Linux, because we set PWD=/proc/self/cwd/, which makes
all paths have a fixed 3-segment prefix. However, this is not possible
on MacOS. On MacOS, the correct setting depends on the path to the exec
root or the sandbox root (if sandboxing is enabled). On my machine, the
proper value is 10.

While this change doesn't fix the underlying issue, at least it allows
users to work around it without having to modify and compile Bazel from
source.

Related to #10457.

PiperOrigin-RevId: 291701562