bazel syntax: remove EvalUtils.getDataTypeName special case for 'select'
This change removes the special case for SelectorList in getDataTypeName
that caused it to return "select of T". Although informative, the element
type is rarely important in the actual errors in which it appears.
To compensate, we work harder to generate good errors when reporting
the wrong kind of select, as in 'select of A + select of B'.

With CL 291262809, this change removes the last obstacle to moving
Selector{List,Value} into lib.packages where they belong.

PiperOrigin-RevId: 291715802