Do not fail or print errors when Shellzelisk cannot find a requested Bazel binary, if tools/bazel exists.
Instead, we should do our best to populate the $BAZEL_REAL env var with some meaningful path, but then delegate to tools/bazel without aborting or printing anything, in case we couldn't find a requested binary or make sense of the contents of .bazelversion.

This is in response to https://github.com/bazelbuild/bazel/issues/10356#issuecomment-576386798 and hopefully fixes the use case reported by the affected users.