bazel syntax: make Location parameters explicitly unused
Many functions accept a Location parameter that was formerly
used to construct EvalExceptions but is now unneeded.
This change removes these where possible (e.g. checkMutable);
those that remain necessary to force overload selection,
such as Dict.put(key, value, loc) are renamed to 'unused'
and documented appropriately.

(A later change will make MutabilityException a subclass of
UnsupportedOperationException, allowing the primary operations
to satisfy the List and Map interfaces in the usual way, just
like ImmutableList. The SkylarkCallable-annotated methods
will use a ForStarlark suffix to avoid the need for overloading.)

Also, move Eval.assignItem to EvalUtils.setIndex, for symmetry
with EvalUtils.index. (It is a primitive operation on values,
not a part of the tree-walking evaluator.)

PiperOrigin-RevId: 291795599