Use a ForkJoinPool for async include scanning
Previously, switching to a FJP was rolled back due to introducing a
performance regression in some cases. While I was unable to reproduce
the regression due to other changes in Bazel, this ties the switch to
the async include scanner flag. I have carefully benchmarked that, and I
think that switching to a FJP is required to roll out the async include
scanner.

PiperOrigin-RevId: 291931477