Use allInputs.size() as size hint for the ActionInputMap.
Without NestedSet-on-Skyframe, getInputDeps returned a map whose size was equal to the number of all inputs of that specific action.
With NestedSet-on-Skyframe, getInputDeps returned a singleton map which was growing over the course of action executions. The resulting size hint was growing with the number of actions being executed (or the total number of inputs), resulting in a ~30% GC tax for big builds.

After this CL, the size hint corresponds again to the expected size of the inputs, also for NestedSet-on-Skyframe.

RELNOTES: None
PiperOrigin-RevId: 291998196