third_party/ijar fix for non-sequentially ordered records in a zip
while trying to smoke test filament 1.4.5 I ran into an issue where ijar was failing on a valid input.  This is because the file entry with offset zero (see bit in bug)
does not appear first.  When ijar encounters this entry, it errantly skips updating
this->p, instead reading from the first byte past the end of the previous file entry.
The entry's size and the header's size don't agree and we fail at zip.cc:395

PiperOrigin-RevId: 292394504